all done

+++ cart.php
@@ -115,20 +115,26 @@ if (isset($_GET['delete_all'])) {
                     $grand_total += $total_amt;
                 }
             } else {
-                echo '<img src = "img/empty-animation1">';
+                echo '
+                    <div class="empty">
+                        <img src="img/empty-animation1.gif" alt="Empty Cart Animation">
+                        <p>No products in your cart yet</p>
+                    </div>
+                ';
             }
+            
 
             ?>
 
         </div>
         <div class="dlt">
-            <a href="cart.php?delete_all" class="btn2">delete all</a>
+            <a href="cart.php?delete_all" class="btn2" onclick="return confirm('do you want to delete all from cart')">delete all</a>
 
         </div>
         <div class="wishlist_total">
             <p>total amount payble : <span>Rs<?php echo $grand_total; ?></span></p>
             <a href="shop.php" class="btn2">continue shopping</a>
-            <a href="checkout.php" class="btn2 <?php echo ($grand_total > 1) ? '' : 'disabled'; ?>" onclick="return confirm('do you want to delete all from wishlist')">proceed to check out</a>
+            <a href="checkout.php" class="btn2 <?php echo ($grand_total > 1) ? '' : 'disabled'; ?>" >proceed to check out</a>
         </div>
     </div>
     <?php include 'footer.php'; ?>
